requirements.txt
beautifulsoup4==4.4.1
bleach==1.5.0
blinker==1.3
boto==2.38.0
chardet==2.3.0
cloud-init==0.7.9
command-not-found==0.3
configobj==5.0.6
cryptography==1.2.3
cycler==0.10.0
decorator==4.1.2
defer==1.0.6
entrypoints==0.2.3
enum34==1.1.6
GDAL==2.2.1
geopy==1.11.0
google-compute-engine==2.4.1
html5lib==0.9999999
idna==2.0
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.1
jedi==0.11.0
Jinja2==2.9.6
jsonpatch==1.10
jsonpointer==1.9
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
language-selector==0.1
lxml==3.5.0
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.7.4
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
notebook==5.1.0
numpy==1.13.3
oauthlib==1.0.3
olefile==0.44
opencv-python==3.3.0.10
pandas==0.20.3
pandocfilters==1.4.2
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
prettytable==0.7.2
prompt-toolkit==1.0.15
protobuf==3.5.1
ptyprocess==0.5.2
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
Pygments==2.2.0
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.2.0
pyserial==3.0.1
pyshp==1.2.12
python-apt==1.1.0b1
python-dateutil==2.6.1
python-debian==0.1.27
python-systemd==231
pytz==2017.2
PyWavelets==0.5.2
pyxdg==0.25
PyYAML==3.11
pyzmq==16.0.2
qtconsole==4.3.1
requests==2.9.1
scikit-image==0.13.1
scikit-learn==0.19.0
scipy==0.19.1
screen-resolution-extra==0.0.0
seaborn==0.8.1
Shapely==1.6.4.post1
simplegeneric==0.8.1
six==1.11.0
ssh-import-id==5.5
system-service==0.3
terminado==0.6
testpath==0.3.1
torch==0.2.0.post3
torchvision==0.1.9
tornado==4.5.2
tqdm==4.19.5
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.13.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.0.3
xkit==0.0.0
